pict: hostile-input hardening + in-CI fuzz suite (decode budget, overflow-safe dims, raw-row fit)
PICT length fields are attacker-controlled; the walker must return Err
— never panic, never allocate unbounded memory — on any byte stream.

New MAX_RASTER_BYTES decode budget (256 MiB): every buffer sized from
hostile fields is checked before allocation — the picFrame canvas (a
12-byte header could previously demand ~17 GB), BitMap / indexed-PixMap
/ DirectBits pixel buffers, and PixPat tiles. All bounds-rectangle
width/height math moves from i16 (debug overflow panic on the legal
-32768..32767 span) to i32 in both decoder and probe.

Fixed a real out-of-bounds panic the suite caught: a 16/32-bit
DirectBitsRect whose rowBytes is smaller than bounds-width x
bytes-per-pixel made write_16bpp_row / write_32bpp_row index past the
row slice. Raw decoders now require the row to fit its pixels (checked
per-decoder — packType 2 legitimately carries 3-byte pixels). Indexed
PixMaps / 1-bpp BitMaps get the equivalent width x bpp <= rowBytes x 8
check + a 1/2/4/8 pixelSize whitelist (shared checked_bitmap_dims).

tests/hostile_round401.rs: every truncation prefix of a 9-stream
opcode-family corpus, seeded xorshift byte mutations, systematic
16-bit length-field maxing at every word offset, and hand-crafted
giant-picFrame / giant-bounds records. Validated locally at 40k
mutations per seed (release) + 15k (debug, overflow checks on).

+++ src/probe.rs
@@ -1029,7 +1029,7 @@ fn skip_raster_opcode_v2(r: &mut Reader<'_>, opcode: u16) -> Result<bool> {
         }
         row_bytes = (rb_raw & 0x3FFF) as usize;
         let bounds = r.read_rect()?;
-        height = (bounds.2 - bounds.0).max(0) as usize;
+        height = (bounds.2 as i32 - bounds.0 as i32).max(0) as usize;
         let _pm_version = r.read_u16()?;
         pack_type = r.read_u16()?;
         let _pack_size = r.read_u32()?;
@@ -1058,7 +1058,7 @@ fn skip_raster_opcode_v2(r: &mut Reader<'_>, opcode: u16) -> Result<bool> {
             indexed_pixmap = true;
             row_bytes = (rb_raw & 0x3FFF) as usize;
             let bounds = r.read_rect()?;
-            height = (bounds.2 - bounds.0).max(0) as usize;
+            height = (bounds.2 as i32 - bounds.0 as i32).max(0) as usize;
             let _pm_version = r.read_u16()?;
             pack_type = r.read_u16()?;
             let _pack_size = r.read_u32()?;
@@ -1087,7 +1087,7 @@ fn skip_raster_opcode_v2(r: &mut Reader<'_>, opcode: u16) -> Result<bool> {
             // Legacy 1-bpp BitMap path.
             row_bytes = rb_raw as usize;
             let bounds = r.read_rect()?;
-            height = (bounds.2 - bounds.0).max(0) as usize;
+            height = (bounds.2 as i32 - bounds.0 as i32).max(0) as usize;
             pack_type = 0;
             pixel_size_for_indexed = 0;
         }
@@ -1192,7 +1192,7 @@ fn skip_pix_pat(r: &mut Reader<'_>) -> Result<()> {
             let rb_raw = r.read_u16()?;
             let row_bytes = (rb_raw & 0x3FFF) as usize;
             let bounds = r.read_rect()?;
-            let height = (bounds.2 - bounds.0).max(0) as usize;
+            let height = (bounds.2 as i32 - bounds.0 as i32).max(0) as usize;
             // pmVersion (2) + packType (2) + packSize (4) + hRes (4) +
             // vRes (4) + pixelType (2) + pixelSize (2) + cmpCount (2) +
             // cmpSize (2) + planeBytes (4) + pmTable (4) +
